Improve install script
Before this commit, the install script was a bit crude and could destroy
the original files under some cirmustances

change the install script to try to be idempotent (try is the key word
there). It will skip if rcfiles if it detects a symlink and will not
overwrite existing files.

add colors. everything looks better in color (except B&W movies)

add message to indicate number of modified files (probably not very
reliable though)